# ai-reply-back-end/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Reply Backend API
  version: "1.0.0"
  description: >
    Мобильный API AI Reply. Тексты сообщений не сохраняются: сервер обрабатывает
    их только в памяти во время запроса.
...
servers:
  - url: https://ai-reply.kz
    description: production
paths:
...
  /api/v1/admin/reports:
    get:
      tags: [admin]
      summary: Жалобы на AI-ответы, новые первыми (identifier — маска почты; text показывается)
      security: [{ adminSession: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, resolved] }, description: "без параметра — все" }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
      responses: { "200": { description: "{ reports: [...], total, page, limit }" }, "400": { description: "INVALID_REQUEST (details.field = status)" } }
  /api/v1/admin/reports/{id}/resolve:
    post:
      tags: [admin]
      summary: Отметить жалобу решённой (аудит report.resolve)
      security: [{ adminSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: "{ ok: true, changed: bool }" }, "404": { description: NOT_FOUND } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminSession:
      type: apiKey
      in: cookie
      name: aireply_admin
      description: Cookie-сессия админ-панели (ADMIN_COOKIE_NAME); для POST нужен ещё заголовок X-CSRF-Token.
...